requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymysql==1.1.1
pyyaml==6.0.2
pydantic==2.10.6
requests==2.32.3
beautifulsoup4==4.12.3